+++ nx/test/nx/defn/evaluator_test.exs
@@ -346,6 +346,23 @@ defmodule Nx.Defn.EvaluatorTest do
       assert tensor == Nx.tensor(3)
     end
 
+    defn metadata_hook(a) do
+      a
+      |> Nx.add(1)
+      |> stop_grad()
+      |> hook(:example, &send_to_self({:default, &1}))
+    end
+
+    test "hook on metadata-wrapped expression" do
+      assert metadata_hook(1) == Nx.tensor(2)
+      assert_received {:default, tensor}
+      assert tensor == Nx.tensor(2)
+
+      assert Nx.Defn.jit(&metadata_hook/1).(1) == Nx.tensor(2)
+      assert_received {:default, tensor}
+      assert tensor == Nx.tensor(2)
+    end
+
     defn container_hook(a, b), do: hook({a, b}, :example, &send_to_self({:default, &1}))
 
     test "container hook with overriddes" do
@@ -684,7 +701,7 @@ defmodule Nx.Defn.EvaluatorTest do
       t = Nx.iota({2, 3}, vectorized_axes: [a: 1], type: :s32)
 
       message = """
-      test/nx/defn/evaluator_test.exs:653: the do-block in while must return tensors with the same shape, type, and names as the initial arguments.
+      test/nx/defn/evaluator_test.exs:670: the do-block in while must return tensors with the same shape, type, and names as the initial arguments.
 
       {\e[32m
        <<<<< Body (do-block) <<<<<
@@ -716,7 +733,7 @@ defmodule Nx.Defn.EvaluatorTest do
 
       error =
         """
-        test/nx/defn/evaluator_test.exs:653: condition must be a scalar tensor, got: #Nx.Tensor<
+        test/nx/defn/evaluator_test.exs:670: condition must be a scalar tensor, got: #Nx.Tensor<
           vectorized[x: 1]
           u8[1]
         \s\s